jacobs base

+++ algorithms/one_hot_gp.py
@@ -23,5 +23,5 @@ class GPOneHotSequenceSpace(GPonRealSpace):
         return μ, var
 
     def predict_f(self, X):
-        μ, var = self.gp.predict_f(tf.constant(X))
+        μ, var = self.gp.predict_f(tf.constant(X.astype(float)))
         return μ, var
